Add repeating redeems feature
The subcommand `redeem` now supports the options `--period` and `--count`.
If both are provided, the redeem will be requested `count` times, with
a delay of `period` milliseconds. There are no checks for failed requests
in this case.

If only one or none of the options are provided, `redeem` acts normally.

(Also add an newline character to the end of `README.md`)